Move document context stuff down to Workspaces
